+++ app/src/main/java/com/yenaly/han1meviewer/ui/screen/home/SubscriptionScreen.kt
@@ -1,85 +1,56 @@
 package com.yenaly.han1meviewer.ui.screen.home
 
-import androidx.compose.animation.AnimatedContent
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.tween
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.togetherWith
 import androidx.compose.foundation.background
-import androidx.compose.foundation.horizontalScroll
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.GridItemSpan
-import androidx.compose.foundation.lazy.grid.LazyGridState
 import androidx.compose.foundation.lazy.grid.rememberLazyGridState
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
-import androidx.compose.material3.Icon
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.LoadingIndicator
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults.pinnedScrollBehavior
 import androidx.compose.material3.pulltorefresh.pullToRefresh
 import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.material3.rememberTopAppBarState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
-import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.platform.LocalWindowInfo
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.yenaly.han1meviewer.R
-import com.yenaly.han1meviewer.logic.model.MySubscriptions
 import com.yenaly.han1meviewer.logic.model.SubscriptionItem
 import com.yenaly.han1meviewer.logic.model.SubscriptionVideosItem
-import com.yenaly.han1meviewer.logic.state.PageLoadingState
 import com.yenaly.han1meviewer.logic.state.WebsiteState
-import com.yenaly.han1meviewer.ui.component.ArtistItem
-import com.yenaly.han1meviewer.ui.component.LoadMoreFooter
 import com.yenaly.han1meviewer.ui.component.PullRefreshOverlay
-import com.yenaly.han1meviewer.ui.component.VideoCardItem
 import com.yenaly.han1meviewer.ui.component.appbar.HanimeScaffold
 import com.yenaly.han1meviewer.ui.component.content.EmptyContent
-import com.yenaly.han1meviewer.ui.component.lazy.LazyVerticalGrid
-import com.yenaly.han1meviewer.ui.preview.fakeArtists
-import com.yenaly.han1meviewer.ui.preview.fakeVideos
-import com.yenaly.han1meviewer.ui.screen.rememberVideoGridColumns
-import com.yenaly.han1meviewer.ui.theme.ArtistIconSize
-import com.yenaly.han1meviewer.ui.theme.SpacingNormal
+import com.yenaly.han1meviewer.ui.screen.home.subscription.SubscriptionContent
+import com.yenaly.han1meviewer.ui.screen.home.subscription.SubscriptionEvent
+import com.yenaly.han1meviewer.ui.screen.home.subscription.SubscriptionUiState
 import com.yenaly.han1meviewer.ui.viewmodel.MySubscriptionsViewModel
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.map
 
+/**
+ * 订阅页面 Screen 层。
+ *
+ * 持有 [MySubscriptionsViewModel]，管理缓存、下拉刷新、加载更多等状态编排。
+ * 渲染委托给 [SubscriptionContent]。
+ *
+ * @param navigateBack 返回回调
+ * @param viewModel 订阅 ViewModel
+ * @param onClickArtist 点击作者 → 跳转搜索
+ * @param onLongClickArtist 长按作者 → 复制分享文本
+ * @param onClickVideosItem 点击视频 → 跳转详情
+ * @param onLongClickVideosItem 长按视频 → 复制分享文本
+ */
 @OptIn(ExperimentalMaterial3ExpressiveApi::class)
 @Composable
-fun SubscriptionApp(
+fun SubscriptionScreen(
     navigateBack: () -> Unit,
     viewModel: MySubscriptionsViewModel,
     onClickArtist: (String) -> Unit,
@@ -89,35 +60,55 @@ fun SubscriptionApp(
 ) {
     val state by viewModel.subscriptionsState.collectAsStateWithLifecycle()
     val cachedArtists = rememberSaveable { mutableStateOf<List<SubscriptionItem>>(emptyList()) }
-    val cachedVideos =
-        rememberSaveable { mutableStateOf<List<SubscriptionVideosItem>>(emptyList()) }
+    val cachedVideos = rememberSaveable { mutableStateOf<List<SubscriptionVideosItem>>(emptyList()) }
     val scrollBehavior = pinnedScrollBehavior(rememberTopAppBarState())
     val gridState = rememberLazyGridState()
 
     val refreshState = rememberPullToRefreshState()
     var isRefreshing by rememberSaveable { mutableStateOf(false) }
-    var expandedItem by rememberSaveable { mutableIntStateOf(-1) }
 
-    if (expandedItem == -1 && cachedVideos.value.isNotEmpty()) {
-        expandedItem = 0
-    }
+    val canLoadMore = viewModel.canLoadMore()
 
-    val onRefresh: () -> Unit = {
-        isRefreshing = true
-        viewModel.loadMySubscriptions(forceReload = true)
-    }
+    val showCached = state is WebsiteState.Loading && cachedArtists.value.isNotEmpty() ||
+            state is WebsiteState.Error && cachedArtists.value.isNotEmpty()
 
     LaunchedEffect(state) {
-        if (state is WebsiteState.Success) {
-            cachedArtists.value =
-                (state as WebsiteState.Success<MySubscriptions>).info.subscriptions.toList()
-            cachedVideos.value =
-                (state as WebsiteState.Success<MySubscriptions>).info.subscriptionsVideos.toList()
-        } else if (state is WebsiteState.Loading && cachedArtists.value.isEmpty()) {
-            viewModel.loadMySubscriptions()
+        when (val s = state) {
+            is WebsiteState.Success -> {
+                cachedArtists.value = s.info.subscriptions.toList()
+                cachedVideos.value = s.info.subscriptionsVideos.toList()
+            }
+            is WebsiteState.Loading -> {
+                if (cachedArtists.value.isEmpty()) viewModel.loadMySubscriptions()
+            }
+            else -> Unit
         }
         if (state !is WebsiteState.Loading) {
             isRefreshing = false
+        }
+    }
+
+    val uiState = SubscriptionUiState(
+        artists = cachedArtists.value,
+        videos = cachedVideos.value,
+        isRefreshing = isRefreshing,
+        canLoadMore = canLoadMore,
+        error = (state as? WebsiteState.Error)?.throwable,
+        showCached = showCached,
+    )
+
+    val handleEvent: (SubscriptionEvent) -> Unit = { event ->
+        when (event) {
+            SubscriptionEvent.OnBack -> navigateBack()
+            is SubscriptionEvent.OnClickArtist -> onClickArtist(event.artistName)
+            is SubscriptionEvent.OnLongClickArtist -> onLongClickArtist(event.artistName)
+            is SubscriptionEvent.OnClickVideo -> onClickVideosItem(event.videoCode)
+            is SubscriptionEvent.OnLongClickVideo -> onLongClickVideosItem(event.videoCode, event.title)
+            SubscriptionEvent.OnRefresh -> {
+                isRefreshing = true
+                viewModel.loadMySubscriptions(forceReload = true)
+            }
+            SubscriptionEvent.OnLoadMore -> viewModel.loadMySubscriptions()
         }
     }
 
@@ -134,26 +125,19 @@ fun SubscriptionApp(
                 .pullToRefresh(
                     state = refreshState,
                     isRefreshing = isRefreshing,
-                    onRefresh = onRefresh
+                    onRefresh = { handleEvent(SubscriptionEvent.OnRefresh) }
                 )
                 .background(MaterialTheme.colorScheme.background)
         ) {
-            when (val result = state) {
+            when (state) {
                 is WebsiteState.Loading -> {
                     if (cachedArtists.value.isEmpty() || cachedVideos.value.isEmpty()) {
                         LoadingIndicator(Modifier.align(Alignment.Center))
                     } else {
-                        // 显示旧数据，刷新中不替换内容
-                        AnimatedPageContent(
+                        SubscriptionContent(
+                            uiState = uiState,
+                            onEvent = handleEvent,
                             gridState = gridState,
-                            artists = cachedArtists.value,
-                            videos = cachedVideos.value,
-                            onClickArtist = onClickArtist,
-                            onLongClickArtist = onLongClickArtist,
-                            onClickVideosItem = onClickVideosItem,
-                            onLongClickVideosItem = onLongClickVideosItem,
-                            onLoadMore = { viewModel.loadMySubscriptions() },
-                            canLoadMore = viewModel.canLoadMore()
                         )
                     }
                 }
@@ -163,38 +147,24 @@ fun SubscriptionApp(
                         EmptyContent(
                             hint = stringResource(
                                 R.string.load_failed_with_reason,
-                                result.throwable.message.orEmpty()
+                                (state as WebsiteState.Error).throwable.message.orEmpty()
                             ),
                             picRes = R.drawable.h_chan_sad
                         )
                     } else {
-                        // 显示旧缓存内容（保持体验）
-                        AnimatedPageContent(
+                        SubscriptionContent(
+                            uiState = uiState,
+                            onEvent = handleEvent,
                             gridState = gridState,
-                            artists = cachedArtists.value,
-                            videos = cachedVideos.value,
-                            onClickArtist = onClickArtist,
-                            onLongClickArtist = onLongClickArtist,
-                            onClickVideosItem = onClickVideosItem,
-                            onLongClickVideosItem = onLongClickVideosItem,
-                            onLoadMore = { viewModel.loadMySubscriptions() },
-                            canLoadMore = viewModel.canLoadMore()
                         )
                     }
                 }
 
-                else -> {
-                    // 统一渲染缓存（成功后缓存已更新）
-                    AnimatedPageContent(
+                is WebsiteState.Success -> {
+                    SubscriptionContent(
+                        uiState = uiState,
+                        onEvent = handleEvent,
                         gridState = gridState,
-                        artists = cachedArtists.value,
-                        videos = cachedVideos.value,
-                        onClickArtist = onClickArtist,
-                        onLongClickArtist = onLongClickArtist,
-                        onClickVideosItem = onClickVideosItem,
-                        onLongClickVideosItem = onLongClickVideosItem,
-                        onLoadMore = { viewModel.loadMySubscriptions() },
-                        canLoadMore = viewModel.canLoadMore()
                     )
                 }
             }
@@ -202,254 +172,6 @@ fun SubscriptionApp(
             PullRefreshOverlay(
                 state = refreshState,
                 isRefreshing = isRefreshing,
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3ExpressiveApi::class)
-@Composable
-fun SubscriptionPageContent(
-    gridState: LazyGridState,
-    videos: List<SubscriptionVideosItem>,
-    artists: List<SubscriptionItem>,
-    onClickVideosItem: (String) -> Unit,
-    onLongClickVideosItem: (String, String) -> Unit,
-    onClickArtist: (String) -> Unit,
-    onLongClickArtist: (String) -> Unit,
-    onLoadMore: () -> Unit,
-    canLoadMore: Boolean
-) {
-    val density = LocalDensity.current
-    val windowInfo = LocalWindowInfo.current
-    val screenWidthPx = windowInfo.containerSize.width
-    val screenWidthDp = with(density) { screenWidthPx.toDp() }
-    val videoColumns = rememberVideoGridColumns()
-    val artistRows = 3
-    val artistColumns = maxOf(3, ((screenWidthDp + SpacingNormal) / (ArtistIconSize + SpacingNormal)).toInt())
-    var currentPage by remember { mutableIntStateOf(1) }
-    val pageSize = 60
-
-    LaunchedEffect(gridState, videos.size) {
-        snapshotFlow { gridState.layoutInfo.visibleItemsInfo }
-            .map { it.lastOrNull()?.index }
-            .distinctUntilChanged()
-            .collect { lastVisibleIndex ->
-                if (lastVisibleIndex != null &&
-                    lastVisibleIndex >= videos.size - 4 &&
-                    videos.size >= currentPage * pageSize &&
-                    canLoadMore
-                ) {
-                    currentPage += 1
-                    onLoadMore()
-                }
-            }
-    }
-
-    LazyVerticalGrid(
-        state = gridState,
-        columns = GridCells.Fixed(videoColumns),
-        modifier = Modifier.fillMaxSize(),
-        contentPadding = PaddingValues(SpacingNormal),
-        horizontalArrangement = Arrangement.spacedBy(SpacingNormal),
-        verticalArrangement = Arrangement.spacedBy(SpacingNormal)
-    ) {
-        // 艺术家
-        item(span = { GridItemSpan(videoColumns) }) {
-            val scrollState = rememberScrollState()
-            val artistColumnCount = maxOf(1, (artists.size + artistRows - 1) / artistRows)
-            val showArtistOverflowHint by remember(scrollState, artistColumnCount, artistColumns) {
-                derivedStateOf {
-                    artistColumnCount > artistColumns && scrollState.value < scrollState.maxValue
-                }
-            }
-            Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 4.dp),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = stringResource(R.string.subscribed_artists_count, artists.size),
-                        style = MaterialTheme.typography.titleSmall,
-                        color = MaterialTheme.colorScheme.onSurface,
-                    )
-                    AnimatedVisibility(
-                        visible = showArtistOverflowHint
-                    ) {
-                        Row(
-                            horizontalArrangement = Arrangement.spacedBy(4.dp),
-                            verticalAlignment = Alignment.CenterVertically,
-                        ) {
-                            Text(
-                                text = stringResource(R.string.swipe_more),
-                                style = MaterialTheme.typography.labelSmall,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                            )
-                            Icon(
-                                painter = androidx.compose.ui.res.painterResource(R.drawable.ic_baseline_arrow_forward_24),
-                                contentDescription = null,
-                                tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                                modifier = Modifier.width(16.dp)
-                            )
-                        }
-                    }
-                }
-                Box(modifier = Modifier.fillMaxWidth()) {
-                    Row(
-                        horizontalArrangement = Arrangement.spacedBy(SpacingNormal),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .horizontalScroll(scrollState)
-                            .padding(end = 28.dp, bottom = 8.dp)
-                    ) {
-                        repeat(artistColumnCount) { columnIndex ->
-                            Column(
-                                modifier = Modifier.width(ArtistIconSize),
-                                verticalArrangement = Arrangement.spacedBy(SpacingNormal)
-                            ) {
-                                repeat(artistRows) { rowIndex ->
-                                    val itemIndex = columnIndex * artistRows + rowIndex
-                                    val artist = artists.getOrNull(itemIndex)
-                                    if (artist != null) {
-                                        ArtistItem(
-                                            artist = artist,
-                                            onClickArtist = onClickArtist,
-                                            onLongClickArtist = onLongClickArtist
-                                        )
-                                    } else {
-                                        Spacer(modifier = Modifier.width(ArtistIconSize))
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    if (showArtistOverflowHint) {
-                        Box(
-                            modifier = Modifier
-                                .align(Alignment.CenterEnd)
-                                .width(32.dp)
-                                .fillMaxHeight()
-                                .background(
-                                    Brush.horizontalGradient(
-                                        colors = listOf(
-                                            Color.Transparent,
-                                            MaterialTheme.colorScheme.background,
-                                        )
-                                    )
-                                )
-                        )
-                    }
-                }
-            }
-        }
-        //分割线
-        item(span = { GridItemSpan(videoColumns) }) {
-            HorizontalDivider(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp),
-                thickness = 1.dp,
-                color = MaterialTheme.colorScheme.outlineVariant
-            )
-        }
-        // 视频列
-        items(videos) { video ->
-            VideoCardItem(
-                videoItem = video,
-                onClickVideosItem = onClickVideosItem,
-                onLongClickVideosItem = {_, _ ->}
-            )
-        }
-        if (videos.isNotEmpty()) {
-            item(span = { GridItemSpan(videoColumns) }) {
-                var loadState by remember {
-                    mutableStateOf<PageLoadingState<*>>(
-                        PageLoadingState.Success(
-                            emptyList<String>()
-                        )
-                    )
-                }
-                LoadMoreFooter(
-                    state = loadState,
-                    isLoadingMore = canLoadMore,
-                    loadedPage = currentPage
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun AnimatedPageContent(
-    gridState: LazyGridState,
-    artists: List<SubscriptionItem>,
-    videos: List<SubscriptionVideosItem>,
-    onClickArtist: (String) -> Unit,
-    onLongClickArtist: (String) -> Unit,
-    onClickVideosItem: (String) -> Unit,
-    onLongClickVideosItem: (String, String) -> Unit,
-    onLoadMore: () -> Unit,
-    canLoadMore: Boolean
-) {
-    AnimatedContent(
-        targetState = Pair(artists, videos),
-        label = "video-content-animation",
-        transitionSpec = {
-            fadeIn(tween(300)) togetherWith fadeOut(tween(200))
-        }
-    ) { (a, v) ->
-        SubscriptionPageContent(
-            gridState = gridState,
-            artists = a,
-            videos = v,
-            onClickArtist = onClickArtist,
-            onLongClickArtist = onLongClickArtist,
-            onClickVideosItem = onClickVideosItem,
-            onLongClickVideosItem = onLongClickVideosItem,
-            onLoadMore = onLoadMore,
-            canLoadMore = canLoadMore
-        )
-    }
-}
-
-@Preview(device = "spec:width=411dp,height=891dp")
-@Composable
-fun SubscriptionAppPreview() {
-    MaterialTheme { SubscriptionAppPreviewBody() }
-}
-
-@Composable
-fun SubscriptionAppPreviewBody() {
-    val scrollBehavior = pinnedScrollBehavior(rememberTopAppBarState())
-    val isRefreshing = false
-    val refreshState = rememberPullToRefreshState()
-
-    HanimeScaffold(
-        modifier = Modifier
-            .nestedScroll(scrollBehavior.nestedScrollConnection)
-            .pullToRefresh(
-                state = refreshState,
-                isRefreshing = isRefreshing,
-                onRefresh = {}
-            ),
-        title = stringResource(R.string.my_subscribe),
-        onBack = {},
-        scrollBehavior = scrollBehavior,
-    ) { innerPadding ->
-        Box(Modifier.padding(innerPadding)) {
-            SubscriptionPageContent(
-                gridState = LazyGridState(),
-                videos = fakeVideos,
-                onClickVideosItem = {},
-                onLoadMore = { },
-                canLoadMore = false,
-                artists = fakeArtists,
-                onClickArtist = {},
-                onLongClickVideosItem = { _, _ -> },
-                onLongClickArtist = { _ -> }
             )
         }
     }
